+++ cookbooks/docker_wrapper/recipes/unifi.rb
@@ -36,14 +36,8 @@ docker_container service do
   hostname "#{node.hostname}-#{service}"
   repo repo
   tag tag
-  restart_policy 'on-failure'
+  restart_policy 'unless-stopped'
   volumes_from service_store
   port ['8080:8080', '8443:8443']
   action :run
-end
-
-## create startup service and script
-docker_wrapper_container_service service do
-  container_name service
-  action :enable
 end